bpf: add hook to parse unknown TCP options
The max option size that we can support for now is 6 bytes, since
the third argument which stores the option value is of u32.
This hook is disabled by default to avoid overhead on the fast path
and are enabled only when BPF_SOCK_OPS_OPTION_PARSE_FLAG is set.