bpf: pass sk struct to tcp_parse_options()
This is necessary for calling TCP-BPF hook since skb->sk is not always set.

We can simplify tcp_parse_options() by removing the net namespace struct
from the argument list. However, it is also called with timewait sock in
tcp_timewait_state_process(). Is it better to push the netns struct
upwards to sock_common since netns is a generic thing?