uto: add mobility & migration scenario with rsync
In a rsync download, client sends UTO option to let old rsync server
to quickly terminate the transfer when the connectivity get lost.

To emulate different servers, set each server iface in a different subnet,
so we set up the middle netns as router instead of bridge.

on client: terminate old connection with rsync's --timeout argument
 (since rsync does not handle TCP timeout nor ICMP dest. unreachable)
and spawn a new rsync client and connect to the new IP address
 (DNS doesn't work, since rsync don't know how to use DNS roundrobin)

on server: change default route to new iface, to emulate a new server