pm: add helper function bpf_open_subflow
Each address need to be passed as a pair of (mem_ptr, arg_size),
as required by the verifier (see commit: 90133415941)
The helper function allows NULL addresses to support ndiffports PM.

We schedule the actual job via work queue, similar to in-kernel PMs.
This is probably because mptcp_init4_subsockets() was designed to be
called from process context, but helper function normally is from
interrupt context.