bpf: fix tcp_parse_options() by passing sk
It would be much easier if skb->sk is always set, but unfortunately it is
not the case.

We can simplify tcp_parse_options() by removing the net namespace struct
from the argument list. However, it is also called with timewait sock in
tcp_timewait_state_process(). Is it better to push the netns struct
upwards to sock_common since netns is a generic thing?
,